Add delete confirmation page for user accounts
Replace the custom page which didn't add any additional value to our shared page

+++ src/AppBundle/Controller/Org/TeamController.php
@@ -213,50 +213,56 @@ class TeamController extends AbstractController
     }
 
     /**
-     * Confirm delete user form
+     * Removes a user, adds a flash message and redirects to page. Asks for confirmation first.
      *
      * @Route("/delete-user/{id}", name="delete_team_member")
-     * @Template()
+     * @Template("AppBundle:Common:confirmDelete.html.twig")
      */
     public function deleteConfirmAction(Request $request, $id, $confirmed = false)
     {
+        $form = $this->createForm(FormDir\ConfirmDeleteType::class);
+        $form->handleRequest($request);
+
+        if ($form->isValid()) {
+            try {
+                $userToRemove = $this->getRestClient()->get('team/member/' . $id, 'User');
+
+                $this->denyAccessUnlessGranted('delete-user', $userToRemove, 'Access denied');
+
+                // delete the user from all the teams the logged user belongs to.
+                // Also removes the user if (after the operation) won't belong to any team any longer
+                $this->getRestClient()->delete('/team/delete-membership/' . $userToRemove->getId());
+
+                $request->getSession()->getFlashBag()->add('notice', 'User account removed');
+            } catch (\Throwable $e) {
+                $this->get('logger')->debug($e->getMessage());
+
+                if ($e instanceof RestClientException && isset($e->getData()['message'])) {
+                    $request->getSession()->getFlashBag()->add(
+                        'error',
+                        'User could not be removed'
+                    );
+                }
+            }
+
+            return $this->redirectToRoute('org_team');
+        }
+
         // The rest call ensures that only team members get returned and permission checks work as expected
         $user = $this->getRestClient()->get('team/member/' . $id, 'User');
 
         $this->denyAccessUnlessGranted('delete-user', $user, 'Access denied');
 
-        return ['user' => $user];
-    }
-
-    /**
-     * Removes a user, adds a flash message and redirects to page
-     *
-     * @Route("/delete-user/{id}/confirm", name="delete_team_member_confirm")
-     * @Template()
-     */
-    public function deleteConfirmedAction(Request $request, $id)
-    {
-        try {
-            $userToRemove = $this->getRestClient()->get('team/member/' . $id, 'User');
-
-            $this->denyAccessUnlessGranted('delete-user', $userToRemove, 'Access denied');
-
-            // delete the user from all the teams the logged user belongs to.
-            // Also removes the user if (after the operation) won't belong to any team any longer
-            $this->getRestClient()->delete('/team/delete-membership/' . $userToRemove->getId());
-
-            $request->getSession()->getFlashBag()->add('notice', 'Operation completed');
-        } catch (\Throwable $e) {
-            $this->get('logger')->debug($e->getMessage());
-
-            if ($e instanceof RestClientException && isset($e->getData()['message'])) {
-                $request->getSession()->getFlashBag()->add(
-                    'error',
-                    'User could not be removed'
-                );
-            }
-        }
-
-        return $this->redirectToRoute('org_team');
+        return [
+            'translationDomain' => 'org-team',
+            'subject' => 'user account',
+            'form' => $form->createView(),
+            'summary' => [
+                ['label' => 'Name', 'value' => $user->getFullName()],
+                ['label' => 'Email address', 'value' => $user->getEmail()],
+                ['label' => 'Administrator?', 'value' => $user->isOrgAdministrator() ? 'Yes' : 'No'],
+            ],
+            'backLink' => $this->generateUrl('org_team'),
+        ];
     }
 }
